requirements.txt
Pillow==10.4.0
PyAutoGUI==0.9.54
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyperclip==1.9.0
requests==2.32.3
pywin32-ctypes==0.2.3
git-changelog==2.5.2
python-dotenv